Fix Neogeo sound crackling (and some FM games), put doubles back from float

+++ src/sound/fmopl.cpp
@@ -708,7 +708,7 @@ static void OPL_initalize(FM_OPL *OPL)
 	/* frequency base */
 	OPL->freqbase = (OPL->rate) ? ((double)OPL->clock / OPL->rate) / 72  : 0;
 	/* Timer base time */
-	OPL->TimerBase = (float)TIME_ONE_SEC/((float)OPL->clock / 72.0 );
+	OPL->TimerBase = (double)TIME_ONE_SEC/((double)OPL->clock / 72.0 );
 	/* make time tables */
 	init_timetables( OPL , OPL_ARRATE , OPL_DRRATE );
 	/* make fnumber -> increment counter table */
